+++ tests/xoto3/dynamodb/write_versioned/transact/ddb_api_test.py
@@ -1,6 +1,9 @@
+import boto3
+import pytest
 from botocore.exceptions import ClientError
 
-from xoto3.dynamodb.write_versioned.ddb_api import is_cancelled_and_retryable
+from xoto3.dynamodb.write_versioned.ddb_api import is_cancelled_and_retryable, known_key_schema
+from xoto3.dynamodb.write_versioned.errors import TableSchemaUnknownError
 
 
 def test_is_cancelled_and_retryable():
@@ -28,3 +31,13 @@ def test_is_cancelled_and_retryable():
             "transact_write_items",
         )
     )
+
+
+def test_key_schema_unfetchable():
+    try:
+        table = boto3.resource("dynamodb").Table("thistabledoesnotexist")
+
+        with pytest.raises(TableSchemaUnknownError):
+            known_key_schema(table)
+    except:  # noqa
+        pass  # test cannot run at all without access to DynamoDB