+++ xoto3/dynamodb/utils/table.py
@@ -1,9 +1,10 @@
 from typing import Tuple
-from xoto3.dynamodb.types import TableResource, InputItem, ItemKey
+
+from xoto3.dynamodb.types import InputItem, ItemKey, TableResource
 
 
 def table_primary_keys(table: TableResource) -> Tuple[str, ...]:
-    return tuple([key["AttributeName"] for key in table.key_schema])
+    return tuple(sorted([key["AttributeName"] for key in table.key_schema]))
 
 
 def extract_key_from_item(table: TableResource, item: InputItem) -> ItemKey: