+++ xoto3/dynamodb/write_versioned/modify.py
@@ -3,14 +3,15 @@
 from typing import NamedTuple, Optional, Union
 
 from xoto3.dynamodb.constants import DEFAULT_ITEM_NAME
-from xoto3.dynamodb.exceptions import get_item_exception_type
 from xoto3.dynamodb.prewrite import dynamodb_prewrite
 from xoto3.dynamodb.types import Item, ItemKey
 from xoto3.utils.tree_map import SimpleTransform
 
+from .ddb_api import known_key_schema
 from .ddb_api import table_name as _table_name
-from .errors import ItemUnknownToTransactionError, TableUnknownToTransactionError
+from .errors import TableSchemaUnknownError
 from .keys import hashable_key, key_from_item
+from .prepare import standard_key_attributes_from_key
 from .types import TableNameOrResource, VersionedTransaction, _TableData
 
 
@@ -32,11 +33,44 @@ def _write(
     *,
     nicename: str = DEFAULT_ITEM_NAME,
 ) -> VersionedTransaction:
-    """Shared put/delete implementation - not meant for direct use at this time."""
+    """Shared put/delete implementation - not meant for direct use at this time.
+
+    Performs an optimistic put - if the item is not known to the
+    existing transaction, assumes you mean to create it if and only if
+    it does not exist. If the item turns out to exist already, your
+    transaction will be re-run, at which point a put will be interpreted as a
+    'witting' choice to overwrite the known item.
+    """
+
     table_name = _table_name(table)
-    if table_name not in transaction.tables:
-        raise TableUnknownToTransactionError(table_name)
-    items, effects, key_attributes = transaction.tables[table_name]
+
+    if table_name in transaction.tables:
+        items, effects, key_attributes = transaction.tables[table_name]
+    else:
+        # we have to have key_attributes in order to proceed with any
+        # effect for the given table. We're going to attempt various
+        # ways of getting them, starting by asking DynamoDB directly.
+        key_attributes = known_key_schema(table)
+        items = dict()
+        effects = dict()
+        if not key_attributes:
+            if isinstance(put_or_delete, Delete):
+                # hope that the user provided an actual item key
+                if len(put_or_delete.item_key) > 2:
+                    raise TableSchemaUnknownError(
+                        f"We don't know the key schema for {table_name} because you haven't defined it "
+                        "and it is not guessable from the delete you requested."
+                        "Specify this delete in terms of the key only and this should work fine."
+                    )
+                # at this point this is a best guess
+                key_attributes = standard_key_attributes_from_key(put_or_delete.item_key)
+            else:
+                # it's a put - we can't make this work at all
+                raise TableSchemaUnknownError(
+                    f"We don't have enough information about table {table_name} to properly derive "
+                    "a key from your request to put this item. "
+                    "Prefetching this item by key would solve that problem."
+                )
 
     item_or_none, item_key = (
         (put_or_delete.item, key_from_item(key_attributes, put_or_delete.item))
@@ -45,13 +79,6 @@ def _write(
     )
 
     hashable_item_key = hashable_key(item_key)
-    if hashable_item_key not in items:
-        # Any item to be modified by the transaction must be specified
-        # at the initiation of the transaction so that it can be
-        # prefetched.
-        raise get_item_exception_type(nicename, ItemUnknownToTransactionError)(
-            "Not specified as part of transaction", key=item_key, table_name=table_name
-        )
 
     return VersionedTransaction(
         tables={