+++ xoto3/dynamodb/write_versioned/read.py
@@ -8,7 +8,7 @@ from xoto3.dynamodb.exceptions import get_item_exception_type, raise_if_empty_ge
 from xoto3.dynamodb.types import Item, ItemKey
 
 from .ddb_api import table_name as _table_name
-from .errors import ItemUnknownToTransactionError, TableUnknownToTransactionError
+from .errors import ItemNotYetFetchedError
 from .keys import hashable_key
 from .types import TableNameOrResource, VersionedTransaction
 
@@ -36,8 +36,9 @@ def get(
     directly. Caveat emptor...
     """
     table_name = _table_name(table)
+    item_exc_type = get_item_exception_type(nicename, ItemNotYetFetchedError)
     if table_name not in transaction.tables:
-        raise TableUnknownToTransactionError(table_name)
+        raise item_exc_type("Table new to transaction", key=item_key, table_name=table_name)
 
     items, effects, _ = transaction.tables[table_name]
     item_hashable_key = hashable_key(item_key)
@@ -46,9 +47,10 @@ def get(
     if item_hashable_key in effects:
         return xf_result(effects[item_hashable_key])
     if item_hashable_key not in items:
-        raise get_item_exception_type(nicename, ItemUnknownToTransactionError)(
-            f"{nicename} not present in transaction", key=item_key, table_name=table_name
+        raise item_exc_type(
+            f"{nicename} not yet present in transaction", key=item_key, table_name=table_name
         )
+
     return xf_result(items[item_hashable_key])